Column sets (#213)
* Do not attach complex default values to prototype

By defining default arrays/objects directly in the component, they will be shared by different instances of the table. This also leads to memory leaks, as the instances are never cleaned up.

* Add columnSets option

This adds the possibility to add columnSets, which are groups of columns that can be toggled together with one click.