add prevent default to `<a>` tag action
The `<a>` tag that calls this action used the `{{on ...}}` modifier. The default action of following the link is triggered and the app navigates to `#` when this tag is clicked. Since this is an `<a>` tag that is actually playing the role of a button I believe we do not want to perform the default action.