Add Volta Config
[Volta](https://volta.sh/) is a JavaScript tool manager that makes it easy to switch between versions of tools like node/yarn/etc between projects. By adding this config, volta users can ensure they are using the expected versions of various tools when working with this project. This has no detrimental effect on non-volta users.

Since the current node range in package.json is backwards compatible to Node 14, I have added the latest Node 14 LTS release in the volta config. When Node 14 is dropped, you can simply update this to the latest Node 16 release, and so on. Additionally, I've pulled in the last yarn 1.x release, so users who have yarn 3 installed don't accidentally overwrite this project's lockfile without the other necessary changes that would be needed to convert to yarn 3.